refactor package to have extensions and differentiate between compiler and runtime
